refactor public profile page
