+++ .meteor/packages
@@ -48,3 +48,4 @@ percolate:migrations
 momentjs:moment
 http
 spiderable
+mrt:headly
