+++ imports/startup/client/routes.js
@@ -25,7 +25,7 @@ import PasswordPage from '/imports/ui/containers/password/PasswordPage';
 import SetPasswordPage from '/imports/ui/containers/password/SetPasswordPage';
 import ForgotAliasPage from '/imports/ui/containers/alias/ForgotAliasPage';
 
-import PublicProfile from '/imports/ui/containers/profile/PublicProfileRefactor';
+import PublicProfile from '/imports/ui/containers/profile/PublicProfile';
 import Profile from '/imports/ui/containers/profile/Profile';
 import Dashboard from '/imports/ui/containers/dashboard/Dashboard';
 import Organizations from '/imports/ui/containers/organizations/Organizations';
